libimage: split out search filters
This allows podman to import them on the remote client without needing
the full libimage.

Based on Miloslav's work: https://github.com/containers/podman/pull/19718
