+++ libimage/search.go
@@ -3,11 +3,10 @@ package libimage
 import (
 	"context"
 	"fmt"
-	"strconv"
 	"strings"
 	"sync"
 
-	"github.com/containers/common/libimage/define"
+	"github.com/containers/common/libimage/filter"
 	registryTransport "github.com/containers/image/v5/docker"
 	"github.com/containers/image/v5/pkg/sysregistriesv2"
 	"github.com/containers/image/v5/transports/alltransports"
@@ -46,7 +45,7 @@ type SearchResult struct {
 // SearchOptions customize searching images.
 type SearchOptions struct {
 	// Filter allows to filter the results.
-	Filter SearchFilter
+	Filter filter.SearchFilter
 	// Limit limits the number of queries per index (default: 25). Must be
 	// greater than 0 to overwrite the default value.
 	Limit int
@@ -75,51 +74,6 @@ type SearchOptions struct {
 	// registry.  If set, the unqualified-search registries in
 	// containers-registries.conf(5) are ignored.
 	Registries []string
-}
-
-// SearchFilter allows filtering images while searching.
-type SearchFilter struct {
-	// Stars describes the minimal amount of starts of an image.
-	Stars int
-	// IsAutomated decides if only images from automated builds are displayed.
-	IsAutomated types.OptionalBool
-	// IsOfficial decides if only official images are displayed.
-	IsOfficial types.OptionalBool
-}
-
-// ParseSearchFilter turns the filter into a SearchFilter that can be used for
-// searching images.
-func ParseSearchFilter(filter []string) (*SearchFilter, error) {
-	sFilter := new(SearchFilter)
-	for _, f := range filter {
-		arr := strings.SplitN(f, "=", 2)
-		switch arr[0] {
-		case define.SearchFilterStars:
-			if len(arr) < 2 {
-				return nil, fmt.Errorf("invalid filter %q, should be stars=<value>", filter)
-			}
-			stars, err := strconv.Atoi(arr[1])
-			if err != nil {
-				return nil, fmt.Errorf("incorrect value type for stars filter: %w", err)
-			}
-			sFilter.Stars = stars
-		case define.SearchFilterAutomated:
-			if len(arr) == 2 && arr[1] == "false" {
-				sFilter.IsAutomated = types.OptionalBoolFalse
-			} else {
-				sFilter.IsAutomated = types.OptionalBoolTrue
-			}
-		case define.SearchFilterOfficial:
-			if len(arr) == 2 && arr[1] == "false" {
-				sFilter.IsOfficial = types.OptionalBoolFalse
-			} else {
-				sFilter.IsOfficial = types.OptionalBoolTrue
-			}
-		default:
-			return nil, fmt.Errorf("invalid filter type %q", f)
-		}
-	}
-	return sFilter, nil
 }
 
 // Search searches term.  If term includes a registry, only this registry will
@@ -261,7 +215,7 @@ func (r *Runtime) searchImageInRegistry(ctx context.Context, term, registry stri
 	paramsArr := []SearchResult{}
 	for i := 0; i < limit; i++ {
 		// Check whether query matches filters
-		if !(options.Filter.matchesAutomatedFilter(results[i]) && options.Filter.matchesOfficialFilter(results[i]) && options.Filter.matchesStarFilter(results[i])) {
+		if !(filterMatchesAutomatedFilter(&options.Filter, results[i]) && filterMatchesOfficialFilter(&options.Filter, results[i]) && filterMatchesStarFilter(&options.Filter, results[i])) {
 			continue
 		}
 		official := ""
@@ -330,18 +284,18 @@ func searchRepositoryTags(ctx context.Context, sys *types.SystemContext, registr
 	return paramsArr, nil
 }
 
-func (f *SearchFilter) matchesStarFilter(result registryTransport.SearchResult) bool {
+func filterMatchesStarFilter(f *filter.SearchFilter, result registryTransport.SearchResult) bool {
 	return result.StarCount >= f.Stars
 }
 
-func (f *SearchFilter) matchesAutomatedFilter(result registryTransport.SearchResult) bool {
+func filterMatchesAutomatedFilter(f *filter.SearchFilter, result registryTransport.SearchResult) bool {
 	if f.IsAutomated != types.OptionalBoolUndefined {
 		return result.IsAutomated == (f.IsAutomated == types.OptionalBoolTrue)
 	}
 	return true
 }
 
-func (f *SearchFilter) matchesOfficialFilter(result registryTransport.SearchResult) bool {
+func filterMatchesOfficialFilter(f *filter.SearchFilter, result registryTransport.SearchResult) bool {
 	if f.IsOfficial != types.OptionalBoolUndefined {
 		return result.IsOfficial == (f.IsOfficial == types.OptionalBoolTrue)
 	}
